+++ src/ContentBlocks/ImageBlock.php
@@ -29,8 +29,6 @@ class ImageBlock extends AbstractFilamentFlexibleContentBlock
     use HasImageConversionType;
     use HasImageWidth;
 
-    public ?string $imageId;
-
     public ?string $imageTitle;
 
     public ?string $imageCopyright;
@@ -44,7 +42,6 @@ class ImageBlock extends AbstractFilamentFlexibleContentBlock
     {
         parent::__construct($record, $blockData);
 
-        $this->imageId = $this->getMediaUuid($blockData['image']) ?? null;
         $this->imageTitle = $blockData['image_title'] ?? null;
         $this->imageCopyright = $blockData['image_copyright'] ?? null;
         $this->imagePosition = $blockData['image_position'] ?? null;
@@ -115,30 +112,21 @@ class ImageBlock extends AbstractFilamentFlexibleContentBlock
 
     public function getImageMedia(?string $conversion = null, array $attributes = []): ?HtmlableMedia
     {
-        return $this->getHtmlableMedia($this->imageId, $this->getImageConversionType($conversion), $this->imageTitle, $attributes);
+        return $this->getHtmlableMedia($this->getBlockId(), $this->getImageConversionType($conversion), $this->imageTitle, $attributes);
     }
 
     public function getImageUrl(?string $conversion = null): ?string
     {
-        return $this->getMediaUrl(imageId: $this->imageId, conversion: $this->getImageConversionType($conversion));
-    }
-
-    public function hasImage(): bool
-    {
-        return isset($this->imageId) && ! is_null($this->imageId);
+        return $this->getMediaUrl(blockId: $this->getBlockId(), conversion: $this->getImageConversionType($conversion));
     }
 
     public function getSearchableContent(): array
     {
         $searchable = [];
 
+        $this->addSearchableContent($searchable, $this->imageTitle);
         $this->addSearchableContent($searchable, $this->imageCopyright);
 
         return $searchable;
-    }
-
-    public function getImageUuids(): array
-    {
-        return $this->imageId ? [$this->imageId] : [];
     }
 }