+++ src/Filament/Form/Fields/TranslatableSpatieMediaLibraryFileUpload.php
@@ -3,13 +3,10 @@
 namespace Statikbe\FilamentFlexibleContentBlocks\Filament\Form\Fields;
 
 use Filament\Forms\Components\SpatieMediaLibraryFileUpload;
-use Illuminate\Database\Eloquent\Model;
 use Illuminate\Support\Collection;
 use Livewire\Component as Livewire;
-use Spatie\MediaLibrary\HasMedia;
 use Spatie\MediaLibrary\MediaCollections\Models\Media;
 use Statikbe\FilamentFlexibleContentBlocks\Filament\Form\Fields\Concerns\HasImageEditor;
-use Throwable;
 
 class TranslatableSpatieMediaLibraryFileUpload extends SpatieMediaLibraryFileUpload
 {
@@ -31,66 +28,6 @@ class TranslatableSpatieMediaLibraryFileUpload extends SpatieMediaLibraryFileUpl
             return $media->filter(function (Media $item) use ($filter) {
                 return $item->getCustomProperty('locale') == $filter['locale'];
             });
-        });
-
-        //temp until PR is merged:
-        $this->loadStateFromRelationshipsUsing(static function (SpatieMediaLibraryFileUpload $component, HasMedia $record): void {
-            /** @var Model&HasMedia $record */
-            $media = $record->load('media')->getMedia($component->getCollection())
-                ->when(
-                    $component->hasMediaFilter(),
-                    fn (Collection $media) => $component->filterMedia($media)
-                )
-                ->when(
-                    ! $component->isMultiple(),
-                    fn (Collection $media): Collection => $media->take(1),
-                )
-                ->mapWithKeys(function (Media $media): array {
-                    $uuid = $media->getAttributeValue('uuid');
-
-                    return [$uuid => $uuid];
-                })
-                ->toArray();
-
-            $component->state($media);
-        });
-
-        //TODO remove after filament release >3.2.72
-        $this->getUploadedFileUsing(static function (SpatieMediaLibraryFileUpload $component, string $file): ?array {
-            if (! $component->getRecord()) {
-                return null;
-            }
-
-            /** @var ?Media $media */
-            $media = $component->getRecord()->getRelationValue('media')->firstWhere('uuid', $file);
-
-            $url = null;
-
-            if ($component->getVisibility() === 'private') {
-                $conversion = $component->getConversion();
-
-                try {
-                    $url = $media?->getTemporaryUrl(
-                        now()->addMinutes(5),
-                        (filled($conversion) && $media->hasGeneratedConversion($conversion)) ? $conversion : '',
-                    );
-                } catch (Throwable $exception) {
-                    // This driver does not support creating temporary URLs.
-                }
-            }
-
-            if ($component->getConversion() && $media?->hasGeneratedConversion($component->getConversion())) {
-                $url ??= $media->getUrl($component->getConversion());
-            }
-
-            $url ??= $media?->getUrl();
-
-            return [
-                'name' => $media?->getAttributeValue('name') ?? $media?->getAttributeValue('file_name'),
-                'size' => $media?->getAttributeValue('size'),
-                'type' => $media?->getAttributeValue('mime_type'),
-                'url' => $url,
-            ];
         });
     }